+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 - Added order weight option for Viewport Input node.
 - Added regular polygons mesh generators menu and search entries.
 - Added curve dimensions option to the Separate Text node.
+- Added matrices output in *Mesh Points Scatter* node.
+- Added edges mode in *Mesh Points Scatter* node.
 
 ### Fixed